tetris: use smaller data types when possible

+++ tetris/include/tetris.h
@@ -9,8 +9,8 @@ enum shape{SquareShape, TShape, ZShape, SShape, JShape, LShape, IShape};
 
 struct pieceInfo{
     uint8_t piece[MAX_PIECE_LEN];
-    uint8_t pieceX;
-    uint8_t pieceY;
+    int8_t pieceX;
+    int8_t pieceY;
     enum shape pieceShape;
 };
 
@@ -26,13 +26,13 @@ void initBoard(uint8_t *boardToInit);
 void initRandomPiece(struct pieceInfo *pieceToInit);
 
 void updateBoard(uint8_t *board, struct pieceInfo *playerPiece);
-void clearRows(uint8_t *boardToClear, int pieceY);
+void clearRows(uint8_t *boardToClear, int8_t pieceY);
 
 enum boardAction handleUserInput(enum userRequest r, uint8_t *board, struct pieceInfo *playerPiece);
 
 bool checkSpawnPiece(uint8_t *pieceToSpawn, uint8_t *boardToSpawnIn, enum shape pieceShape);
 void renderBoard(uint8_t *renderedBoard, uint8_t *boardToRender, struct pieceInfo *pieceToRender, bool substract);
-bool checkMove(uint8_t *piece, uint8_t piecePosX, uint8_t piecePosY, enum shape pieceShape, uint8_t *boardToCheck);
+bool checkMove(uint8_t *piece, int8_t piecePosX, int8_t piecePosY, enum shape pieceShape, uint8_t *boardToCheck);
 
 
 #endif
